+++ Casso65Emu/Pch.h
@@ -5,6 +5,7 @@
 #include <windows.h>
 
 #include <d3d11.h>
+#include <d3dcompiler.h>
 #include <dxgi.h>
 #include <mmdeviceapi.h>
 #include <audioclient.h>